requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.0.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
numpy==1.16.2
pkg-resources==0.0.0
PyYAML==5.1
typing-extensions==3.7.2
websockets==6.0
yarl==1.3.0
